requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.30.0
aiohttp~=3.9.5
httpx~=0.27.0
pydantic~=2.7.1
dnspython~=2.6.1
python-dotenv~=1.0.1